stock manage

+++ src/product/product.service.ts
@@ -1,9 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UseGuards } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from './entities/product.entity';
 import { DataSource, Repository } from 'typeorm';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { RolesGuard } from 'src/auth/roles.guard';
+import { Roles } from 'src/auth/decorators/roles.decorator';
 
 
 @Injectable()
@@ -15,6 +18,8 @@ export class ProductService {
     private readonly dataSource: DataSource
   ) { }
 
+  @UseGuards(AuthGuard,RolesGuard)
+  @Roles('ADMIN')
   async create(createProductDto: CreateProductDto) {
     const data = this.productRepository.create(createProductDto);
     return this.productRepository.save(data);
@@ -25,19 +30,22 @@ export class ProductService {
     limit = Number(limit) || 10;
     const result = this.productRepository
       .createQueryBuilder("product")
-      .leftJoin("product.brand", "b")
-      .leftJoin("product.category", "c")
-      .leftJoin("product.colors", "pc")
-      .leftJoin("product.labels", "pl")
+      .innerJoin('product.variants','pv')
+      .leftJoin('pv.label','pl')
+      .leftJoin('pv.color','pc')
+      .leftJoin('product.brand','b')
+      .leftJoin('product.category','c')
       .select([
-        "product.quantity AS Quantity",
+        'product.id as ID',
+        "pv.stock AS Quantity",
         "product.name AS ProductName",
-        "product.price AS ProductPrice",
-        "b.name AS BrandName",
-        "c.name AS CategoryName",
-        "pc.value AS AvailableColor",
-        "pc.hex_value AS ColorHexValue",
-        "pl.value AS LabelValue",
+        "pv.price AS ProductPrice",
+        "pc.value as Color",
+        "pc.hex_value as HexColorCode",
+        "pl.value as Label",
+        "b.name  as Brand",
+        'c.name as Category'
+        
       ])
       .limit(limit)
       .offset(skip)
@@ -58,24 +66,25 @@ export class ProductService {
     const results =
       await this.productRepository
         .createQueryBuilder("p")
+        .innerJoin('p.variants','pv')
         .leftJoin('p.brand', 'b')
         .leftJoin('p.category', 'c')
-        .leftJoin("p.colors", "pc")
-        .leftJoin("p.labels", "pl")
-        .leftJoin("p.orders_items", "orders")
+        .leftJoin("pv.color", "pc")
+        .leftJoin("pv.label", "pl")
+        
         .select([
-          "p.id",
-          "p.quantity AS Quantity",
+          "p.id as Id",
+          "pv.stock as AvailableQuantity",
+          "pv.price as Price",
           "p.name AS ProductName",
-          "p.price AS ProductPrice",
+
           "b.name AS brandName",
           "c.name AS categoryName",
           "pc.value AS colorValue",
           "pc.hex_value AS ColorHexValue",
           "pl.value AS labelValue",
-          "orders.name AS orderItemName"
         ])
-        .where("p.id = :id", { id: id }).getRawOne()
+        .where("p.id = :id", { id: id }).getRawMany()
     return results;
   }
 
@@ -87,13 +96,20 @@ export class ProductService {
 
 
   async update(id: number, updateProductDto: UpdateProductDto) {
-    const find: any = await this.findOne(id);
+    const find: any = await this.productRepository.findOneByOrFail({id})
+    if(!find){
+      return `Product Not Found`
+    }
     const data = this.productRepository.merge(find, updateProductDto)
     return this.productRepository.save(data)
   }
 
   async remove(id: number) {
-    const find: any = await this.findOne(id);
+    const find: any = await this.productRepository.findOneByOrFail({id})
+    console.log(find)
+    if(!find){
+      return `Product Not Found`
+    }
     return this.productRepository.remove(find);
   }
 }
